backend/requirements.txt
fastapi
python-multipart
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
httpx
pypdf
python-docx
